完成用户管理和聊天室的功能

+++ SiteMgr/schSocialMedia/views.py
@@ -1,30 +1,102 @@
 from django.http import HttpResponse
-from .models import ArticleComments, Article
+from .models import ArticleComments, Article, User
 from django.shortcuts import render, redirect
-
-from .models import ArticleComments, Article, Room
-from .forms import RoomForm
+from django.db.models import Q
+from .models import ArticleComments, Article, Room, Topic, Message
+from .forms import RoomForm, CustomUserCreationForm
+from django.contrib import messages
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth.forms import UserCreationForm
 
 
 # Create your views here.
+
+def loginPage(request):
+
+    page = 'login'
+    if request.user.is_authenticated:
+        return redirect('index')
+
+    if request.method =='POST':
+        username = request.POST.get('username').lower()
+        password = request.POST.get('password')
+
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request, '用户不存在')
+
+        user= authenticate(request,username=username,password=password)
+
+        if user is not None:
+            login(request, user)
+            return redirect('index')
+        else:
+            messages.error(request, '用户或密码不正确')
+
+    context = {'page': page}
+    return render(request, 'schSocialMedia/login_register.html', context)
+
+
+def logoutUser(request):
+    logout(request)
+    return  redirect('index')
+
+def registerPage(request):
+    page ='register'
+    form = CustomUserCreationForm()
+    if request.method == 'POST':
+        form = CustomUserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request,user)
+            return redirect('index')
+        else:
+            messages.error(request, "注册出现异常")
+    context = {'page': page,'form': form}
+    return render(request, 'schSocialMedia/login_register.html', context)
 
 def index(request):
     # atcl = Article.objects.get(id=1)
     # cmts = atcl.articlecomments_set.all()[1]
     # atclcmts = ArticleComments.objects.filter(article=atcl)
     # return HttpResponse(atclcmts[0])
-    rooms = Room.objects.all()
-    content = {'rooms': rooms}
+    q = request.GET.get('q') if request.GET.get('q') != None else ''
+    rooms = Room.objects.filter(Q(topic__name__icontains=q) |
+                                Q(name__icontains=q) |
+                                Q(content__icontains=q)
+                                )
+
+    room_count = rooms.count()
+    topics = Topic.objects.all()
+
+    content = {'rooms': rooms, 'topics': topics, 'room_count': room_count}
     return render(request, "schSocialMedia/index.html", content)
     pass
 
 
 def room(request, pk):
     room = Room.objects.get(id=pk)
-    content = {'room': room}
+    room_messages = room.message_set.all().order_by('-pub_date')
+    participants = room.participants.all()
+
+    if request.method == 'POST':
+        message = Message.objects.create(
+            user=request.user,
+            room=room,
+            content=request.POST.get('body')
+        )
+        room.participants.add(request.user)
+        return redirect('room', pk=room.id)
+
+    content = {'room': room,'room_messages': room_messages,'participants': participants}
     return render(request, "schSocialMedia/room.html", content)
 
 
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -37,10 +109,15 @@ def createRoom(request):
     context = {'form': form}
     return render(request, "schSocialMedia/room_form.html", context)
 
+@login_required(login_url='login')
 def updateRoom(request, pk):
 
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        messages.error(request, '不能操作别人的房间')
+        return redirect('index')
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -51,10 +128,30 @@ def updateRoom(request, pk):
     context = {'form': form}
     return render(request, "schSocialMedia/room_form.html", context)
 
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room =Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        messages.error(request, '不能操作别人的房间')
+        return redirect('index')
+
     if request.method == 'POST':
         room.delete()
         return redirect('index')
 
     return render(request, "schSocialMedia/delete.html", {'obj': room})
+
+@login_required(login_url='login')
+def deleteMessage(request, pk):
+    message =Message.objects.get(id=pk)
+    room = message.room
+    if request.user != message.user:
+        messages.error(request, '不能操作别人的房间')
+        return redirect('index')
+
+    if request.method == 'POST':
+        message.delete()
+        return redirect('room', pk=room.id)
+
+    return render(request, "schSocialMedia/delete.html", {'obj': message})
